refactor(checker): split conformance issue regression tests
